Remove transactions folder and move extraction logic to the programs
