+++ crates/contracts/src/lib.rs
@@ -1,4 +1,3 @@
 pub mod artifacts;
 pub mod programs;
-pub mod transactions;
 pub mod utils;
