+++ crates/contracts/src/programs/issuance_factory/error.rs
@@ -1,3 +1,5 @@
+use simplex::simplicityhl::elements::Txid;
+
 #[derive(thiserror::Error, Debug)]
 pub enum IssuanceFactoryError {
     #[error("Invalid creation OP_RETURN data length: expected - {expected}, actual - {actual}")]
@@ -5,4 +7,10 @@ pub enum IssuanceFactoryError {
 
     #[error("Invalid OP_RETURN owner pubkey bytes: {0}")]
     InvalidOpReturnBytes(String),
+
+    #[error("Confidential assets currently are not supported")]
+    ConfidentialAssetsAreNotSupported(),
+
+    #[error("Passed transaction is not an issuance factory creation transaction")]
+    NotAnIssuanceFactoryCreationTx(Txid),
 }
