+++ crates/contracts/src/programs/lending/error.rs
@@ -1,0 +1,10 @@
+use simplex::simplicityhl::elements::Txid;
+
+#[derive(thiserror::Error, Debug)]
+pub enum LendingError {
+    #[error("Confidential assets currently are not supported")]
+    ConfidentialAssetsAreNotSupported(),
+
+    #[error("Passed transaction is not a lending creation transaction")]
+    NotALendingCreationTx(Txid),
+}
