+++ crates/contracts/src/programs/lending/mod.rs
@@ -1,7 +1,9 @@
 mod core;
+mod error;
 mod params;
 mod witness;
 
 pub use core::Lending;
+pub use error::LendingError;
 pub use params::LendingParameters;
 pub use witness::LendingWitnessBranch;
