+++ crates/contracts/src/programs/pre_lock/core.rs
@@ -1,16 +1,18 @@
 use simplex::program::Program;
 
-use simplex::provider::SimplicityNetwork;
-use simplex::simplicityhl::elements::hex::ToHex;
-use simplex::simplicityhl::elements::secp256k1_zkp::XOnlyPublicKey;
-use simplex::simplicityhl::elements::{AssetId, Script};
+use simplex::provider::{ProviderTrait, SimplicityNetwork};
+use simplex::simplicityhl::elements::{
+    AssetId, Script, Transaction, hex::ToHex, secp256k1_zkp::XOnlyPublicKey,
+};
 use simplex::transaction::{FinalTransaction, PartialOutput, UTXO};
+use simplex::utils::hash_script;
 
 use crate::artifacts::pre_lock::PreLockProgram;
 use crate::programs::lending::Lending;
-use crate::programs::pre_lock::{PreLockParameters, PreLockWitnessBranch, error::PreLockError};
+use crate::programs::pre_lock::{PreLockError, PreLockParameters, PreLockWitnessBranch};
 use crate::programs::program::SimplexProgram;
 use crate::programs::script_auth::{ScriptAuth, ScriptAuthWitnessParams};
+use crate::utils::{FirstNFTParameters, LendingOfferParameters, SecondNFTParameters};
 
 pub struct PreLock {
     program: PreLockProgram,
@@ -25,6 +27,85 @@ impl PreLock {
             program: PreLockProgram::new(parameters.build_arguments()),
             parameters,
         }
+    }
+
+    pub fn try_from_tx(
+        tx: &Transaction,
+        provider: &impl ProviderTrait,
+    ) -> Result<Self, PreLockError> {
+        if tx.input.len() < 5 || tx.output.len() < 7 || !tx.output[5].is_null_data() {
+            return Err(PreLockError::NotAPreLockCreationTx(tx.txid()));
+        }
+
+        let collateral_asset_id = tx.output[0]
+            .asset
+            .explicit()
+            .ok_or_else(PreLockError::ConfidentialAssetsAreNotSupported)?;
+        let first_parameters_nft_asset_id = tx.output[1]
+            .asset
+            .explicit()
+            .expect("Utility NFT must be explicit");
+        let second_parameters_nft_asset_id = tx.output[2]
+            .asset
+            .explicit()
+            .expect("Utility NFT must be explicit");
+        let borrower_nft_asset_id = tx.output[3]
+            .asset
+            .explicit()
+            .expect("Utility NFT must be explicit");
+        let lender_nft_asset_id = tx.output[4]
+            .asset
+            .explicit()
+            .expect("Utility NFT must be explicit");
+
+        let first_parameters_nft_amount = tx.output[1]
+            .value
+            .explicit()
+            .expect("Parameter NFT must have explicit amount");
+        let second_parameters_nft_amount = tx.output[2]
+            .value
+            .explicit()
+            .expect("Parameter NFT must have explicit amount");
+
+        let offer_parameters = LendingOfferParameters::build_from_parameters_nfts(
+            &FirstNFTParameters::decode(first_parameters_nft_amount),
+            &SecondNFTParameters::decode(second_parameters_nft_amount),
+        );
+
+        let prev_collateral_outpoint = tx.input[0].previous_output;
+        let pre_collateral_tx = provider.fetch_transaction(&prev_collateral_outpoint.txid)?;
+        let collateral_script_hash = hash_script(
+            &pre_collateral_tx.output[prev_collateral_outpoint.vout as usize].script_pubkey,
+        );
+
+        let mut op_return_instr_iter = tx.output[5].script_pubkey.instructions_minimal();
+
+        op_return_instr_iter.next();
+
+        let op_return_bytes = op_return_instr_iter
+            .next()
+            .unwrap()
+            .unwrap()
+            .push_bytes()
+            .unwrap();
+
+        let (borrower_pubkey, principal_asset_id) =
+            PreLock::decode_creation_op_return_data(op_return_bytes.to_vec()).unwrap();
+
+        let pre_lock_parameters = PreLockParameters {
+            collateral_asset_id,
+            principal_asset_id,
+            first_parameters_nft_asset_id,
+            second_parameters_nft_asset_id,
+            borrower_nft_asset_id,
+            lender_nft_asset_id,
+            offer_parameters,
+            borrower_pubkey,
+            borrower_output_script_hash: collateral_script_hash,
+            network: *provider.get_network(),
+        };
+
+        Ok(Self::new(pre_lock_parameters))
     }
 
     pub fn get_parameters(&self) -> &PreLockParameters {
