+++ crates/contracts/src/programs/pre_lock/mod.rs
@@ -4,5 +4,6 @@ mod params;
 mod witness;
 
 pub use core::PreLock;
+pub use error::PreLockError;
 pub use params::PreLockParameters;
 pub use witness::PreLockWitnessBranch;
