--- crates/contracts/src/transactions/core.rs
@@ -1,53 +1,0 @@
-use simplex::simplicityhl::elements::{AssetId, Script};
-use simplex::transaction::{PartialInput, PartialOutput, RequiredSignature, UTXO};
-
-#[derive(Debug, Clone)]
-pub struct SimplexInput {
-    partial_input: PartialInput,
-    required_sig: RequiredSignature,
-}
-
-impl SimplexInput {
-    pub fn new(utxo: &UTXO, required_sig: RequiredSignature) -> Self {
-        Self {
-            partial_input: PartialInput::new(utxo.clone()),
-            required_sig,
-        }
-    }
-
-    pub fn explicit_asset(&self) -> AssetId {
-        self.partial_input.asset.expect("Not an explicit asset")
-    }
-
-    pub fn explicit_amount(&self) -> u64 {
-        self.partial_input.amount.expect("Not an explicit amount")
-    }
-
-    pub fn utxo_script_pubkey(&self) -> Script {
-        self.partial_input.witness_utxo.script_pubkey.clone()
-    }
-
-    pub fn new_partial_output(&self) -> PartialOutput {
-        PartialOutput::new(
-            self.utxo_script_pubkey(),
-            self.explicit_amount(),
-            self.explicit_asset(),
-        )
-    }
-
-    pub fn new_burn_partial_output(&self) -> PartialOutput {
-        PartialOutput::new(
-            Script::new_op_return(b"burn"),
-            self.explicit_amount(),
-            self.explicit_asset(),
-        )
-    }
-
-    pub fn partial_input(&self) -> &PartialInput {
-        &self.partial_input
-    }
-
-    pub fn required_sig(&self) -> &RequiredSignature {
-        &self.required_sig
-    }
-}
