--- crates/contracts/src/transactions/lending/error.rs
@@ -1,16 +1,0 @@
-use simplex::simplicityhl::elements::Txid;
-
-#[derive(thiserror::Error, Debug)]
-pub enum LendingTransactionError {
-    #[error("Confidential assets currently are not supported")]
-    ConfidentialAssetsAreNotSupported(),
-
-    #[error("Not enough principal assets to repay: expected - {expected}, actual - {actual}")]
-    NotEnoughPrincipalToRepay { expected: u64, actual: u64 },
-
-    #[error("Passed transaction is not a lending creation transaction")]
-    NotALendingCreationTx(Txid),
-
-    #[error("Failed to convert loan expiration time to LockTime: {0}")]
-    InvalidLockHeight(u32),
-}
