--- crates/contracts/src/transactions/lending/mod.rs
@@ -1,5 +1,0 @@
-pub mod error;
-pub mod parameters;
-
-pub use error::*;
-pub use parameters::*;
