--- crates/contracts/src/transactions/mod.rs
@@ -1,5 +1,0 @@
-pub mod core;
-// pub mod issuance_factory;
-// pub mod lending;
-// pub mod pre_lock;
-pub mod utility;
