--- crates/contracts/src/transactions/utility/mod.rs
@@ -1,7 +1,0 @@
-pub mod error;
-pub mod issuance_preparation;
-pub mod utility_nfts_issuance;
-
-pub use error::*;
-pub use issuance_preparation::*;
-pub use utility_nfts_issuance::*;
